requirements.txt
langchain==0.1.20
openai==1.30.1
faiss-cpu==1.7.4
gradio==4.26.0
pypdf==4.2.0
python-dotenv==1.0.1 